:truck: refactor(api): move session summaries to `src`
