+++ api/lib/application/certification-centers/certification-center-controller.js
@@ -5,20 +5,6 @@ import { usecases } from '../../domain/usecases/index.js';
 import * as sessionSummarySerializer from '../../infrastructure/serializers/jsonapi/session-summary-serializer.js';
 import * as studentCertificationSerializer from '../../infrastructure/serializers/jsonapi/student-certification-serializer.js';
 
-const findPaginatedSessionSummaries = async function (request) {
-  const certificationCenterId = request.params.id;
-  const userId = request.auth.credentials.userId;
-  const options = request.query;
-
-  const { models: sessionSummaries, meta } = await usecases.findPaginatedCertificationCenterSessionSummaries({
-    userId,
-    certificationCenterId,
-    page: options.page,
-  });
-
-  return sessionSummarySerializer.serialize(sessionSummaries, meta);
-};
-
 const getStudents = async function (request) {
   const certificationCenterId = request.params.certificationCenterId;
   const sessionId = request.params.sessionId;
@@ -91,7 +77,6 @@ const updateReferer = async function (request, h) {
 const certificationCenterController = {
   createCertificationCenterMembershipByEmail,
   findCertificationCenterMembershipsByCertificationCenter,
-  findPaginatedSessionSummaries,
   getDivisions,
   getStudents,
   updateReferer,
