+++ api/lib/application/certification-centers/index.js
@@ -149,29 +149,6 @@ const register = async function (server) {
         tags: ['api', 'certification-center', 'students', 'session'],
       },
     },
-    {
-      method: 'GET',
-      path: '/api/certification-centers/{id}/session-summaries',
-      config: {
-        validate: {
-          params: Joi.object({
-            id: identifiersType.certificationCenterId,
-          }),
-          query: Joi.object({
-            page: Joi.object({
-              number: Joi.number().integer().empty('').allow(null).optional(),
-              size: Joi.number().integer().empty('').allow(null).optional(),
-            }).default({}),
-          }),
-        },
-        handler: certificationCenterController.findPaginatedSessionSummaries,
-        tags: ['api', 'certification-center'],
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n',
-          '- Elle retourne les sessions rattachées au centre de certification.',
-        ],
-      },
-    },
   ]);
 };
