+++ api/tests/unit/application/certification-centers/index_test.js
@@ -1,5 +1,5 @@
 import { certificationCenterController } from '../../../../lib/application/certification-centers/certification-center-controller.js';
-import * as moduleUnderTest from '../../../../lib/application/certification-centers/index.js';
+import * as moduleUnderTest from '../../../../src/certification/session-management/application/certification-centers-session-summaries.route.js';
 import { securityPreHandlers } from '../../../../src/shared/application/security-pre-handlers.js';
 import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
 
@@ -270,22 +270,6 @@ describe('Unit | Router | certification-center-router', function () {
 
       // then
       expect(result.statusCode).to.equal(400);
-    });
-  });
-
-  describe('GET /api/certification-centers/{certificationCenterId}/session-summaries', function () {
-    it('should return 200', async function () {
-      // given
-      sinon.stub(certificationCenterController, 'findPaginatedSessionSummaries').returns('ok');
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/certification-centers/123/session-summaries');
-
-      // then
-      expect(response.statusCode).to.equal(200);
-      sinon.assert.calledOnce(certificationCenterController.findPaginatedSessionSummaries);
     });
   });
 });
